requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.1
getenv==0.2.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.19.0
